\name{commspacetime-package}
\alias{commspacetime-package}
\alias{commspacetime}
\docType{package}
\title{
\packageTitle{commspacetime}
}
\description{
\packageDescription{commspacetime}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{commspacetime}
\packageIndices{commspacetime}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{commspacetime}

Maintainer: \packageMaintainer{commspacetime}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
