% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorption.R
\name{find_absorption}
\alias{find_absorption}
\title{Absorption function}
\usage{
find_absorption(f, temp, C = TRUE, rh, p = 101.325)
}
\arguments{
\item{f}{Frequency of vocalization (Hz)}

\item{temp}{Temperature}

\item{C}{If your temperature is in Celsius, mark TRUE. If Fahrenheit, mark FALSE.}

\item{rh}{Relative humidity (percentage)}

\item{p}{Barometric pressure. Set to 101.325 as a default.}
}
\value{
Atmospheric absorption--this gets multiplied by the radius in that calculation.
}
\description{
If the vocalizations are far enough away that atmospheric absorption
might become an issue, use this function to calculate it to increase
the accuracy of your radius calculation.
}
\examples{
find_absorption(f = 2250, temp = 10, C = T, rh = 75, p)
}

