% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_calc.R
\name{find_r}
\alias{find_r}
\title{Radius calculation}
\usage{
find_r(signal, noise, r_vector, absorption = 1, perception = 0)
}
\arguments{
\item{signal}{The source level (dB re 20 microPa) of the vocalization.}

\item{noise}{The ambient noise level (dB re 20 microPa) of the place where the animal is vocalizing.}

\item{r_vector}{The initialized vector of possible radius values.}

\item{absorption}{The atmospheric absorption calculated by the absorption.R function. Set to 1 unless otherwise specified.}

\item{perception}{The detection threshold for your species in question, if known. By default, this is set to zero, which indicates the best possible scenario in which any signal can be perceived if it is above the level of background noise.}
}
\description{
This function calculates the radius of the sphere or hemisphere of 
communication. This can then be used in future calculations.
It requires an initialized vector to work with--the upper limit of this
can be determined by your species' vocalization propoerties.
}
\examples{
find_r(90, 65, r_vector, absorption = 1, perception)
}

