% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_volume.R
\name{time_volume}
\alias{time_volume}
\title{Time-volume function}
\usage{
time_volume(duration, rate, radius, sphere = T)
}
\arguments{
\item{duration}{How long each vocalization lasts for (seconds)}

\item{rate}{How often each vocalization is produced per minute}

\item{radius}{Calculated from the radius_calc.r function--radius of the sphere}

\item{sphere}{If the vocalization is produced above the ground, this is set to true. If it's produced on the ground, it's set to false.}
}
\description{
The last function in the series of modeling commmunication space-time.
}
\examples{
time_volume(0.02, 37, 19, sphere = F)
}

